Kickoff
Game is on.

Aim of this commit is to start version control.
Tests are green, but it doesn't mean you can use it out of the box.
It doesn't have a box.

Currently it can authenticate and request player by id or
all players for given team by id.

I have worked hard to scrape player data from html, but it still
does not work completely.

But there are 2 tests that should work if user-secrets are setup.

TODO:
 - Refactor, refactor, refactor.
 - I think that I don't like any class, method or prop name.
 - I need to separate html scraping, authentication and
   domain concepts.
 - I haven't set for architecture yet. Currently it is state-driven,
   but I need to consider event-driven benefits.
 - I need to figure out somekind of UI.

Good luck to me.
Siim Haas